refactor: auth 관련 파일을 features에서 entities로 이동 및 경로 수정
