+++ src/entities/auth/services/LoginResponse.ts
