+++ src/entities/auth/services/getTokens.ts
@@ -1,6 +1,6 @@
 import { CognitoUserSession } from "amazon-cognito-identity-js";
 
-import getLoggedIn from "./getLoggedIn";
+import getLoggedIn from "@/entities/auth/services/getLoggedIn";
 
 export interface UserTokens {
     idToken: string;
