+++ src/pages/auth/index.tsx
@@ -2,10 +2,10 @@ import React from "react";
 
 import { Flex, Stack } from "@chakra-ui/react";
 
-import AuthControlsPresentation from "@/features/auth/ui/AuthControlsPresentation";
-import TokenInfoPresentation from "@/pages/auth/ui/TokenInfoPresentation";
-import UserInfoPresentation from "@/features/auth/ui/UserInfoPresentation";
-import getLoggedIn from "@/features/auth/lib/getLoggedIn";
+import AuthControls from "@/features/auth/ui/AuthControls";
+import TokenInfo from "@/pages/auth/ui/TokenInfo";
+import UserInfo from "@/features/userInfo/ui/UserInfo";
+import getLoggedIn from "@/entities/auth/services/getLoggedIn";
 
 export default function AuthPage(): React.ReactElement {
     const { isLoggedIn } = getLoggedIn();
@@ -13,11 +13,11 @@ export default function AuthPage(): React.ReactElement {
     return (
         <Flex bg="gray.100" p={3}>
             {!isLoggedIn ? (
-                <AuthControlsPresentation />
+                <AuthControls />
             ) : (
                 <Stack>
-                    <UserInfoPresentation />
-                    <TokenInfoPresentation />
+                    <UserInfo />
+                    <TokenInfo />
                 </Stack>
             )}
         </Flex>
